ingest/product_crawler/requirements.txt
scrapy>=2.8.0
pymongo
pyyaml